remove unused code, log unmount rc, require hash for http mirrors
- drop unused enum members, dataclass fields, methods, yaml keys
  across config/builder/data/gui
- move PartitionBar from gui/tabs/partitions.py to gui/widgets/
- log rc on macos diskutil unmount + windows set-disk offline
- read kickstart rom bytes once for hash + header parse
- skip http aminet mirrors when no expected_hash is set
- restrict github 'tag' field charset to ^[\w.\-+]+$
- collapse 4 multi-line docstrings, drop 'we' in one
