+++ src/main/python/emu68hatcher/gui/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""custom Qt widgets reused across tabs"""
